requirements.txt
# HTTP and API clients
requests>=2.31.0
aiohttp>=3.9.0
websockets>=12.0

# GraphQL client
gql[all]>=3.5.0

# CLI framework
click>=8.1.7
typer>=0.9.0

# Terminal UI
rich>=13.7.0

# Configuration
toml>=0.10.2

# Data handling
pandas>=2.1.0
python-dateutil>=2.8.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
responses>=0.24.1

# Notifications
plyer>=2.1.0

